+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change GE Proton to be installed during the first-time setup.
 - Change the GRUB theme from Steam Big Picture to Vimix.
 - Add pre-built AUR packages to the winesapOS repository.
+- Change from 'pipewire-media-session' to 'wireplumber' for modern PipeWire support.
 
 ## 3.1.0 - The Controllers and Flatpak Update